requirements.txt
PyHamcrest==2.1.0
pytest==8.3.4
requests==2.32.3